Refactor backend package structure
